MUC: split ChatWindow into private and MUC
